LatexilaSynctex
It is needed for the build tools for opening a PDF file with synctex
support with evince. Since I don't want to use the Vala code in C, a
rewrite of synctex.vala in C was needed.

synctex.vala uses the synchronous D-Bus API. The rewrite in C uses the
asynchronous API (better).

TODO: send a signal for the sync_source (PDF -> .tex)