Add completion menu item (ctrl+space)
The Ctrl+space shortcut was documented nowhere in the interface, only in
the documentation. Now with the menu item, the Ctrl+space shortcut is
more discoverable.

Several years ago there was a bug with the "fr oss" keyboard layout: a
space (without the Ctrl modifier pressed) triggered Ctrl+space, see
https://bugzilla.gnome.org/show_bug.cgi?id=611089
It seems to be fixed now.