Release 3.14.0 - Pi version!
It's easier to follow GNOME versions. For the packagers, for the users,
for the contributors, and for the maintainers. When creating a
latexila-x-y branch, a mail must be manually sent to the GNOME
translation and documenation teams, which is not convenient. On the
other hand, when creating a gnome-x-y branch, mails are sent
automatically.

The 2.0 version was chosen because there was a major code change (a
complete rewrite). But now only incremental changes are made every six
months.