Change Help menu item for the fundraiser
I prefer "LaTeXila Fundraiser" than "LaTeXila Finance", it's more explicit.

Don't translate the strings since the dialog window is in English, so it
is more consistent. I prefer that the dialog window is in English
because if I want to change the content later, the problem with
half-translated content is avoided (and for a fundraiser campaign it's
important...). It'd be possible to untranslate all paragraphs directly
by changing the *.po files, but it isn't convenient, and there is
another reason, the web page for the fundraiser is also in English and
isn't translated. So it's simpler to keep all fundraiser-related text in
English.