TemplatesDefault and TemplatesPersonal classes
Slightly better code design.

Before, LatexilaTemplates (the model) contained functions for the view,
returning a GtkTreeView. Then for some functions the GtkTreePath must
come from the good view; just the fact that it must be documented is a
smell for a bad design. Now a GtkTreePath argument to a GtkListStore
subclass makes perfect sense, and doesn't need extra documentation.