DeleteTemplateDialog: delete the good templates
This is an important bug fix. Wrong templates could have been deleted!
This bug is present in latexila since many versions, and hasn't been
reported (the FIXME comment has been added recently). I guess the dialog
window is usually used to delete only one template at a time (in that
case the bug doesn't occur). Many users are hopefully not aware that it
is possible to select several templates and delete them in one go.