Personal templates: store filenames in the RC file
The purpose is to have less IO operations when deleting a template or
moving a template up or down.

This is backward compatible: templates created with an earlier version
of latexila will still work fine. But if templates are deleted or moved
with latexila >= 3.18, coming back to an earlier version of latexila
will not work.