TemplatesPersonal: ensure that the RC file is always loaded
When creating a new template when the object was not already completely
initialized, it tried to create the new template with the name 0.tex,
which may be already used.

So, load the RC file synchronously, to be sure that the templates are
loaded. This is not a nice solution, I would prefer to do every I/O
operation asynchronously to not block the UI. Hopefully a later commit
will fix that. In the meantime, this is a quicker and simpler solution.