spell: menu entry affects only current tab
The "highlight misspelled words" menu entry now affects only the current
tab. If we use the inline checker to spell check several files, it is
useful to know where the spell checking is finished, i.e. where the
inline checker is disabled. It's just one example, there are probably
many other examples where the inline checker should be activated on a
file-by-file basis.

A later commit will store the spell checking settings to the file
metadata.