Add an app menu
Better late than never. It's redundant with the menubar, but it's better
to keep the items in the menubar too because the app menu is difficult
to discover. The app menu is just to make GNOME people happy (but IIRC
I've never received any complaints about the lack of app menu in
LaTeXila).