app: fix new-window action when launching primary instance
From gnome-shell, if latexila is present in the favorites, doing a right
click -> 'Open a New Window' works fine now. If latexila was not already
launched, only one window is created (not two). And if latexila was
already launched, a new window is created.

It also works when doing a right click -> 'Open a New Document'.

Or when opening files from Nautilus.