app: update copyright in the about dialog
Remove the ©, it's redundant with the word "Copyright", and © is
sometimes not displayed nicely (too small).

Is it the last copyright update? Will latexila die? Will latexila find a
new maintainer? Or will I still be motivated to do the minimum
maintenance work required to still use latexila with the latest GTK+ 3.x
version? Will latexila be ported to GTK+ 4 in the future? Was latexila
just a real-world experiment that permitted me to better learn
programming? Will Vala still be maintained in 10 years? So many
uncertainties.