Add release dates to the NEWS file
I think it's useful information. It can be retrieved from git, but to do
it correctly (get the commiter date of the commit that the git tag
points to) is more difficult. It's easier if it's present in the NEWS
file.

The dates have been added with this quick and dirty script:
```
	#!/usr/bin/env bash

	tags_list=`git tag`
	for tag in $tags_list
	do
		# For the date, it takes the commiter date of the git commit that $tag
		# points to. This is normally the closest date to the real date of the
		# release (if the tag refers to a version). Sometimes a maintainer
		# forgets to create the tag and the tag is created afterwards.
		#
		# To show git tags alongside the tag creation dates, use the following
		# command instead:
		# $ git tag -l --format="%(refname:strip=2) - %(taggerdate:short)"

		commiter_date=`git log -1 --date=short --pretty="format:%cd" $tag`

		sed -i "s/^\(Release ${tag#v}\)$/\1, ${commiter_date}/" NEWS
	done

	sed -i "s/\(--------\)/\1------------/" NEWS
```

For the versions <= 0.2, the dates have been added manually from the
information in the first git repo (no longer public).