Document: weak ref on the DocumentTab
If latexila opens just one DocumentTab, on application exit there were 3
remaining refs to that DocumentTab. Now it's reduced to 2 remaining
refs. So the memory leak is not fixed, but it's one step closer to fix
it.

Note that an object should not be aware of who contains it. It's fine
for DocumentTab to know about Document, since Document is contained in
DocumentTab, but the reverse is not a good practice. This code was
written a long time ago…

https://bugzilla.gnome.org/show_bug.cgi?id=761181