Call tepl_application_window_set_tab_group()
For read purposes, the core of the Tepl framework is now complete! Write
purposes will come later (e.g. to add a new tab).

But with read purposes (e.g. get the list of TeplTab's or get the active
TeplTab in a window), it's already quite powerful. It'll be possible to
implement a lot of actions (as GActions) without depending on the
LaTeXila Vala code. For example I think that it is now possible to
rewrite latex_menu.vala in C (at least implementing the actions as
GActions).