Document: use tepl_file_get_short_name()
When the location is null, the TeplFile:short-name is set synchronously.
On the other hand when the location isn't null, the :short-name is
updated asynchronously inside TeplFile, so it's dangerous to call
get_short_name() without listening to notify::short-name.

So use get_short_name() only when location is null, as a first step.