MainWindow: rework sensitivity updates
This fixes a recent regression introduced by
commit bcff602c95208b167249cfd596d038583dbf7a1b:

For the build tools actions: empty notebook -> open a file -> the build
tools actions were still insensitive. This is now fixed. The
MainWindowBuildTools's update_sensitivity() function was called too
early, the active-tab property was still null.

So now, update sensitivity (and do other stuff) when the
notify::active-tab signal is sent by the TeplApplicationWindow, so we
are sure that the active-tab property has the new value.