MainWindow: connect to notify::active-tab at the end of initialization
When the signal is sent, all latexila widgets/objects must have been
created (at least those used in the signal handler). To be sure that all
objects are created, connect to the signal at the end of the MainWindow
initialization. And set_tab_group() after connecting to the signals, so
if the notebook already contains a tab, the signal handlers will be
called.

This commit has normally no visible effect in latexila, it is just to
have safer and more future-proof code.